Let the DefaultBuildContext delegate to the legacy build-api
Currently there is a problem that if a maven-plugin wants to upgrade to
the newer API artifact it looses backward-compatibility to older
implementors of the API instantly.

This changes the DefaultBuildContext in a way that allows it to behave
backward-compatible in this case:

1) it gets injected the old implementation
2) it delegates all relevant calls to the legacy
3) it contains a feature-switch that is able to detect if the
default-legacy-api is used and therefore we can exchange behavior with a
new default or need to still to delegate to a custom implementation.