Use EventSpy to handle execution of events
Currently we use an AbstractMavenLifecycleParticipant but this only
offers limited number of events, an EventSpy is much more flexible and
offers much more insights.

This now replaces the SessionListener with an EventSpy and also adds
support for project execution events. Alongside with that the
configuration is now local to the EventListener and no longer part of
the connection implementation.

Signed-off-by: Christoph Läubrich <christoph@laeubi-soft.de>